Refactor: Update CrewAI usage to modern syntax
- Update to modern CrewAI syntax using @CrewBase decorator
- Create standalone tool with @tool decorator
- Move tool implementation out of the class
- Add YAML configuration files
- Fix tool integration issues
